+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - [#165](https://github.com/LayerManager/layman/issues/165) Many endpoints respect role access rights:
   - [GET](doc/rest.md#get-workspace-layer) Workspace Layer
   - [GET](doc/rest.md#get-workspace-map) Workspace Map
+  - GET Workspace [Layers](doc/rest.md#get-workspace-layers)/[Maps](doc/rest.md#get-workspace-maps)
+  - GET [Layers](doc/rest.md#get-layers)/[Maps](doc/rest.md#get-maps)/[Publications](doc/rest.md#get-publications)
 - [#165](https://github.com/LayerManager/layman/issues/165) Many endpoints return previously associated [role names](doc/models.md#role) in `access_rights.read` and `access_rights.write` keys:
   - [GET](doc/rest.md#get-workspace-layer)/[PATCH](doc/rest.md#patch-workspace-layer) Workspace Layer
   - [GET](doc/rest.md#get-workspace-map)/[PATCH](doc/rest.md#patch-workspace-map) Workspace Map